# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: LedgerSync Private API
  version: 3.3.0
  description: >-
    Closed-loop single-currency ledger API behind the LedgerSync BFF. Exact
    minor units and monotonic versions are JSON strings. A timed-out transfer
    has an unknown outcome and must be retried with the identical payload and
    Idempotency-Key. Account and controlled-funding command timeouts follow the
    same retry rule. Funding records external value evidence; it does not claim
    bank settlement, custody, or movement outside this ledger.
  license: { name: Apache License 2.0, identifier: Apache-2.0 }
servers: [{ url: /api }]
security: [{ bearerAuth: [], actorAssertion: [] }]
...
paths:
...
  /transfers:
    get:
      operationId: listTransfers
      x-required-scope: transfers:read
      summary: Search tenant transfer outcomes.
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { in: query, name: accountId, schema: { type: string, format: uuid } }
        - { in: query, name: status, description: 'Server-side financial-status filter applied before pagination.', schema: { type: string, enum: [pending, posted, rejected] } }
        - { in: query, name: q, description: 'Server-side case-insensitive substring search over transfer, source-account, and destination-account UUIDs; hexadecimal and hyphen characters only. The cursor is bound to this and every other filter.', schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[0-9A-Fa-f-]+$' } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        '200': { description: Stable descending transfer page., content: { application/json: { schema: { $ref: '#/components/schemas/TransferPage' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: createTransfer
      x-required-scope: transfers:write
      summary: Submit an exact internal transfer under tenant policy.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationID' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTransferRequest' } } }
      responses:
        '201':
          description: Final posted or rejected idempotent outcome.
          headers:
            Idempotent-Replay: { schema: { type: string, enum: ['true'] } }
            X-LedgerSync-Consistency-Requirements: { schema: { type: string } }
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferResult' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/PolicyDenied' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /transfers/{transferId}:
    get:
      operationId: getTransfer
      x-required-scope: transfers:read
      summary: Inspect a transfer, postings, and evidence timeline.
      parameters: [{ $ref: '#/components/parameters/TransferID' }]
      responses:
        '200': { description: Transfer detail., content: { application/json: { schema: { $ref: '#/components/schemas/TransferDetail' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /reconciliation/runs:
    get:
      operationId: listReconciliationRuns
      x-required-scope: reconciliation:read
      summary: List persisted reconciliation runs.
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { description: Stable reconciliation page., content: { application/json: { schema: { $ref: '#/components/schemas/ReconciliationPage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      operationId: runReconciliation
      x-required-scope: reconciliation:write
      summary: Run one tenant-wide reconciliation under transaction-safe single-run exclusion.
      description: >-
        Requires reconciliation:write. The request has no body and requires an
        Idempotency-Key. Matched and mismatch are both durable completed run
        outcomes; already_running is a stable, idempotently replayable denial.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationID' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, maxProperties: 0, additionalProperties: false } } }
      responses:
        '201':
          description: Original or safely replayed matched/mismatch reconciliation result.
          headers: { Idempotent-Replay: { schema: { type: string, enum: ['true'] } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ReconciliationRun' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ReconciliationConflict' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ReconciliationUnavailable' }
        '504': { $ref: '#/components/responses/ReconciliationResponseUnknown' }
...
  /reconciliation/runs/{runId}:
    get:
      operationId: getReconciliationRun
      x-required-scope: reconciliation:read
      summary: Read a run and immutable mismatch evidence.
      parameters: [{ $ref: '#/components/parameters/RunID' }]
      responses:
        '200': { description: Reconciliation detail., content: { application/json: { schema: { $ref: '#/components/schemas/ReconciliationRun' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    actorAssertion:
      type: apiKey
      in: header
      name: X-LedgerSync-Actor-Assertion
      description: Short-lived BFF-signed end-user context bound to the workload token.
  parameters:
...
    TransferID: { in: path, name: transferId, required: true, schema: { type: string, format: uuid } }
...
    RunID: { in: path, name: runId, required: true, schema: { type: string, format: uuid } }
...
    Cursor: { in: query, name: cursor, schema: { type: string, maxLength: 512 } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
...
    IdempotencyKey: { in: header, name: Idempotency-Key, required: true, schema: { type: string, minLength: 16, maxLength: 255 } }
    CorrelationID: { in: header, name: X-Request-ID, schema: { type: string, format: uuid } }
...
  schemas:
    ExactMinor: { type: string, pattern: '^-?(0|[1-9][0-9]*)$' }
    ExactVersion: { type: string, pattern: '^(0|[1-9][0-9]*)$' }
    Currency: { type: string, pattern: '^[A-Z]{3}$', example: INR }
...
    CreateTransferRequest:
      type: object
      required: [source_account_id, destination_account_id, amount, currency]
      additionalProperties: false
      properties:
        source_account_id: { type: string, format: uuid }
        destination_account_id: { type: string, format: uuid }
        amount: { type: string, pattern: '^[0-9]+(\.[0-9]+)?$', example: '125.50' }
        currency: { $ref: '#/components/schemas/Currency' }
    TransferResult:
      type: object
      required: [transfer_id, status, currency, amount_minor, occurred_at, minimum_balance_versions]
      properties:
        transfer_id: { type: string, format: uuid }
        status: { type: string, enum: [posted, rejected] }
        currency: { $ref: '#/components/schemas/Currency' }
        amount_minor: { $ref: '#/components/schemas/ExactMinor' }
        occurred_at: { type: string, format: date-time }
        minimum_balance_versions: { type: object, description: "Read-your-writes versions only for accounts the initiating actor may read.", additionalProperties: { $ref: '#/components/schemas/ExactVersion' } }
        balances: { type: object, description: "Committed balances only for accounts the initiating actor may read; credit permission alone never discloses a destination balance.", additionalProperties: { $ref: '#/components/schemas/TransferBalance' } }
        rejection_code: { type: string }
    TransferBalance:
      type: object
      required: [account_id, currency, posted_minor, version, as_of]
      properties:
        account_id: { type: string, format: uuid }
        currency: { $ref: '#/components/schemas/Currency' }
        posted_minor: { $ref: '#/components/schemas/ExactMinor' }
        version: { $ref: '#/components/schemas/ExactVersion' }
        as_of: { type: string, format: date-time }
...
    TransferSummary:
      type: object
      required: [transfer_id, source_account_id, destination_account_id, amount_minor, currency, financial_status, delivery_status, created_at, completed_at]
      properties:
        transfer_id: { type: string, format: uuid }
        source_account_id: { type: string, format: uuid }
        destination_account_id: { type: string, format: uuid }
        amount_minor: { $ref: '#/components/schemas/ExactMinor' }
        currency: { $ref: '#/components/schemas/Currency' }
        financial_status: { type: string, enum: [pending, posted, rejected] }
        delivery_status: { type: string, enum: [not_applicable, pending, retrying, delivered, dead] }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        journal_transaction_id: { type: string, format: uuid }
        rejection_code: { type: string }
        correction_id: { type: string, format: uuid }
        correction_status: { type: string, enum: [requested, approved, rejected, cancelled, expired, posted] }
        correction_role: { type: string, enum: [original, compensation] }
        original_transfer_id: { type: string, format: uuid }
        compensation_transfer_id: { type: string, format: uuid }
        original_journal_id: { type: string, format: uuid }
        compensation_journal_id: { type: string, format: uuid }
    Posting:
      type: object
      required: [posting_id, account_id, direction, amount_minor, currency, occurred_at]
      properties:
        posting_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        direction: { type: string, enum: [debit, credit] }
        amount_minor: { $ref: '#/components/schemas/ExactMinor' }
        currency: { $ref: '#/components/schemas/Currency' }
        occurred_at: { type: string, format: date-time }
    TransferDetail:
      allOf:
        - { $ref: '#/components/schemas/TransferSummary' }
        - type: object
          required: [actor_subject_id, postings, timeline]
          properties:
            actor_subject_id: { type: string }
            postings: { type: array, items: { $ref: '#/components/schemas/Posting' } }
            timeline: { type: array, items: { type: object } }
...
    TransferPage: { type: object, required: [transfers, next_cursor], properties: { transfers: { type: array, items: { $ref: '#/components/schemas/TransferSummary' } }, next_cursor: { type: string } } }
    ReconciliationMismatch:
      type: object
      required: [mismatch_id, classification, created_at]
      properties:
        mismatch_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        classification: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        expected_minor: { $ref: '#/components/schemas/ExactMinor' }
        observed_minor: { $ref: '#/components/schemas/ExactMinor' }
        observed_available_minor: { $ref: '#/components/schemas/ExactMinor' }
        balance_version: { $ref: '#/components/schemas/ExactVersion' }
        created_at: { type: string, format: date-time }
    ReconciliationRun:
      type: object
      required: [run_id, status, correlation_id, scope, ledger_watermark, application_version, schema_version, checked_account_count, posting_count, mismatch_count, started_at, completed_at]
      properties:
        run_id: { type: string, format: uuid }
        status: { type: string, enum: [running, matched, mismatch, failed] }
        correlation_id: { type: string, format: uuid }
        scope: { type: string }
        ledger_watermark: { type: string }
        application_version: { type: string }
        schema_version: { type: string }
        checked_account_count: { $ref: '#/components/schemas/ExactVersion' }
        posting_count: { $ref: '#/components/schemas/ExactVersion' }
        mismatch_count: { $ref: '#/components/schemas/ExactVersion' }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        mismatches: { type: array, items: { $ref: '#/components/schemas/ReconciliationMismatch' } }
    ReconciliationPage: { type: object, required: [runs, next_cursor], properties: { runs: { type: array, items: { $ref: '#/components/schemas/ReconciliationRun' } }, next_cursor: { type: string } } }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, format: uuid }
  examples:
...
    ReadOnlyDenial: { value: { error: { code: forbidden, message: You cannot perform this action., request_id: 00000000-0000-4000-8000-000000000099 } } }
...
    ReconciliationAlreadyRunning: { value: { error: { code: reconciliation_already_running, message: A reconciliation run is already active for this tenant., request_id: 00000000-0000-4000-8000-000000000099 }, run_id: 00000000-0000-4000-8000-000000000077 } }
    ReconciliationResponseUnknown: { value: { error: { code: response_unknown, message: The reconciliation outcome is unknown. Retry with the identical idempotency key., request_id: 00000000-0000-4000-8000-000000000099 } } }
  responses:
    InvalidRequest: { description: Validation failed without a financial change., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json; media-type parameters such as charset are allowed., content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: { code: unsupported_media_type, message: Content-Type must be application/json., request_id: 00000000-0000-4000-8000-000000000099 } } } } }
    Unauthorized: { description: Authentication is required., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Access denied without resource disclosure., content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, examples: { readOnlyDenial: { $ref: '#/components/examples/ReadOnlyDenial' } } } } }
    NotFound: { description: Authorized resource not found., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Idempotency conflict or in-progress request., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    ReconciliationConflict: { description: 'A run is already active, the key belongs to another intent, or the same request is still completing. already_running includes the active run_id when one remains active.', headers: { Idempotent-Replay: { schema: { type: string, enum: ['true'] } } }, content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/Error' }, { type: object, properties: { run_id: { type: string, format: uuid } } }] }, examples: { alreadyRunning: { $ref: '#/components/examples/ReconciliationAlreadyRunning' } } } } }
    ReconciliationUnavailable: { description: Reconciliation could not start safely; retry the identical intent and idempotency key., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    ReconciliationResponseUnknown: { description: The caller timed out after submission and must retry the identical intent and idempotency key., content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, examples: { responseUnknown: { $ref: '#/components/examples/ReconciliationResponseUnknown' } } } } }
...
    PolicyDenied: { description: "Tenant amount, velocity, destination, or currency policy denied the transfer.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited:
      description: Principal/tenant/route limit exceeded.
      headers: { Retry-After: { required: true, schema: { type: integer, minimum: 1 } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unavailable: { description: Authoritative evidence or current balance unavailable., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    InternalError: { description: Non-sensitive server error., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }